Change prop type of Amount to decimal

+++ expensetrackerapi/DTO/UpdateTransactionDto.cs
@@ -14,7 +14,7 @@ public class UpdateTransactionDto
 
     [Required(ErrorMessage = "The Amount is required.")]
     [Range(0, 100_000, ErrorMessage = "The amount must be between 0 and 100_000.")]
-    public required int Amount { get; set; }
+    public required decimal Amount { get; set; }
 
     [CreatedAtValidation]
     [Required(ErrorMessage = "Created_at is required.")]
